move some block create utils into scene_creator/blocks/
