+++ pgdrive/scene_creator/blocks/t_intersection.py
@@ -1,10 +1,10 @@
 from pgdrive.pg_config.parameter_space import Parameter, BlockParameterSpace
 from pgdrive.pg_config.pg_space import PgSpace
 from pgdrive.scene_creator.basic_utils import Goal
+from pgdrive.scene_creator.blocks.create_block_utils import CreateRoadFrom
 from pgdrive.scene_creator.blocks.intersection import InterSection
 from pgdrive.scene_creator.lanes.lane import LineType, LineColor
 from pgdrive.scene_creator.road.road import Road
-from pgdrive.scene_creator.basic_utils import CreateRoadFrom
 
 
 class TInterSection(InterSection):
